+++ packages/ui/react-ui-graph/src/components/styles.ts
@@ -1,0 +1,8 @@
+//
+// Copyright 2022 DXOS.org
+//
+
+/**
+ * NOTE: Import default styles from @dxos/react-ui-graph/styles.
+ */
+export const defaultStyles = 'graph'; // TODO(burdon): Remove.